Codebase/requirements.txt
torch
transformers
sentencepiece
sentence-transformers
Pillow
reportlab
